Refactor: Remove QueueSong class and consolidate into Song class
• Removed QueueSong class entirely
• Moved all QueueSong properties and methods to Song class:
  - SavedPosition, CurrentPosition, State, WasPlaying
  - AddedToQueueAt, StartedPlayingAt, PausedAt, TotalPlayTime
  - SaveCurrentPosition(), RestorePosition(), RecordPlaybackStart(), etc.
• Added DownloadInfo property to Song class for download events
• Updated Queue class to work directly with Song instances
• Updated QueueCacheService to use Song instead of QueueSong
• Updated MusicPlayerService to remove QueueSong type checks
• Renamed PlayQueueSongAsync to PlaySongWithPositionRestorationAsync
• Removed GetCurrentQueueSong() method
• All download events can now pass the whole Song instance around
