+++ StreamingPlayerNET/Services/QueueCacheService.cs
@@ -61,9 +61,7 @@ public class QueueCacheService
                 // Create queue cache data
                 var queueCache = new QueueCacheData
                 {
-                    Songs = queueToSave.Songs.Take(config.MaxCachedQueueSize)
-                        .Select(s => s is QueueSong ? (QueueSong)s : QueueSong.FromSong(s))
-                        .ToList(),
+                    Songs = queueToSave.Songs.Take(config.MaxCachedQueueSize).ToList(),
                     CurrentIndex = queueToSave.CurrentIndex,
                     RepeatMode = queueToSave.RepeatMode,
                     ShuffleEnabled = queueToSave.ShuffleEnabled,
@@ -116,9 +114,7 @@ public class QueueCacheService
                 // Create queue cache data
                 var queueCache = new QueueCacheData
                 {
-                    Songs = queue.Songs.Take(config.MaxCachedQueueSize)
-                        .Select(s => s is QueueSong ? (QueueSong)s : QueueSong.FromSong(s))
-                        .ToList(),
+                    Songs = queue.Songs.Take(config.MaxCachedQueueSize).ToList(),
                     CurrentIndex = queue.CurrentIndex,
                     RepeatMode = queue.RepeatMode,
                     ShuffleEnabled = queue.ShuffleEnabled,
@@ -200,7 +196,7 @@ public class QueueCacheService
 
 public class QueueCacheData
 {
-    public List<QueueSong> Songs { get; set; } = new();
+    public List<Song> Songs { get; set; } = new();
     public int CurrentIndex { get; set; } = -1;
     public RepeatMode RepeatMode { get; set; } = RepeatMode.None;
     public bool ShuffleEnabled { get; set; } = false;
